feature: Add `NewObjectArray` w/ rule impl
This adds the `NewObjectArray` struct for inferring array data resulting
from invocations of the `New-Object` cmdlet. It handles cases where the
type is either `byte[]` or `int[]`, and includes a constructor allowing
for a `max_args` parameter to prevent OOM DoS via an excessively large
size parameter in the script text.